pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "HiClass"
version = "0.1.0"
description = "Hierarchy Classification Tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat==0.5.7",
    "autogen-core==0.5.7",
    "autogen-ext[anthropic,docker]==0.5.7",
    "black>=25.1.0",
    "blackdoc>=0.3.9",
    "docker==7.1.0",
    "isort>=6.0.1",
    "pytest==8.3.5",
    "tiktoken>=0.9.0",
    "pytest-asyncio==0.26.0",
    "pymupdf==1.26.3",
    "python-dotenv==1.1.1",
    "openai>=1.100.2",
]

[tool.setuptools]
py-modules = ["HiClass", "prompt"]